The brain extraction tool is now testing for the "sphericity" of the connected components it extracts to select the brain. Before this improvement, it indeed happened in about 10% of the cases that the skull formed a larger connected component than the brain and was selected instead.
The method works as follows. For each component, a sphere of 25 mm radius is placed at the center of gravity of 
the component. The proportion of that sphere that is included in the component is then computed. 
That allows to reliably select the brain :
 * large portion of the skull/scalp have a center of gravity which is usually located outside of the component. In any 
case, those structures are thin and won't include more than half of a 5 cm diameter ball
 * the brain, on the other hand, often includes 95% of such a sphere, even without filling the ventricles.