Implemented external subtypes
There are still a number of assumptions about these subtypes. For example, we expect them to be ordered in ascending order and to be named network_i where `i` is the number of the network. The only thing we currently check is that there are enough networks to operate on.